requirements.txt
openai>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
jieba>=0.42.1
regex>=2023.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
loguru>=0.7.0